+++ client/src/middlewares/pagination.rs
@@ -1,4 +1,4 @@
-pub fn limit(count: u32, page: u32) -> String {
+pub fn limit(count: usize, page: usize) -> String {
     let offset = if page > 0 { page * count } else { 0 };
     format!("limit={}&offset={}", count, offset)
 }
